Move scrambleLocation to utils/geo
