+++ src/components/ClickableMap/ClickableMap.js
@@ -1,11 +1,12 @@
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import Geocode from 'react-geocode';
 import {
   GoogleMap,
   LoadScript,
   InfoWindow,
   Marker,
 } from '@react-google-maps/api';
-import PropTypes from 'prop-types';
-import React, { useState } from 'react';
 import {
   Typography,
   Button,
@@ -14,40 +15,12 @@ import {
   makeStyles,
 } from '@material-ui/core';
 import { useNotifications } from 'modules/notification';
-import Geocode from 'react-geocode';
+import { scrambleLocation } from 'utils/geo';
 import styles from './ClickableMap.styles';
 
 const useStyles = makeStyles(styles);
-Geocode.setApiKey(process.env.REACT_APP_GOOGLE_MAPS_API_KEY);
-Geocode.setRegion('us');
 
-// Based on https://stackoverflow.com/a/31280435/391230
-function scrambleLocation(center, radiusInMeters) {
-  const y0 = center.lat;
-  const x0 = center.lng;
-  const rd = radiusInMeters / 111300; // about 111300 meters in one degree
-
-  const u = Math.random();
-  const v = Math.random();
-
-  const w = rd * Math.sqrt(u);
-  const t = 2 * Math.PI * v;
-  const x = w * Math.cos(t);
-  const y = w * Math.sin(t);
-
-  // Adjust the x-coordinate for the shrinking of the east-west distances
-  const xp = x / Math.cos(y0);
-
-  const newlat = y + y0;
-  const newlon2 = xp + x0;
-
-  return {
-    _latitude: Math.round(newlat * 1e5) / 1e5,
-    _longitude: Math.round(newlon2 * 1e5) / 1e5,
-  };
-}
-
-function ClickableMap(props) {
+function ClickableMap({ onLocationChange }) {
   const classes = useStyles();
   const [map, setMap] = useState(null);
   const { showSuccess, showError } = useNotifications();
@@ -55,21 +28,22 @@ function ClickableMap(props) {
   const [markerLocation, setMarkerLocation] = useState(null);
   const [generalLocationName, setGeneralLocationName] = useState('');
 
-  const setLocation = async (location) => {
-    // console.log("called", location);
+  async function setLocation(location) {
+    // console.log('setLocation called:', location);
     if (map) {
-      // console.log("center", location);
-      // map.panTo(location);
-      // TODO: Figure out how to properly zoom and pan.
-      // if (map.getZoom() < 10) {
-      //   map.setZoom(10);
-      // }
+      //   console.log("center", location);
+      //   map.panTo(location);
+      //   // TODO: Figure out how to properly zoom and pan.
+      //   if (map.getZoom() < 10) {
+      //     map.setZoom(10);
+      //   }
     }
 
     const scrambledLocation = scrambleLocation(location, 300); // Roughly within 1,000 feet.
-
-    // Detect locality.
+    // Detect locality
     try {
+      Geocode.setApiKey(process.env.REACT_APP_GOOGLE_MAPS_API_KEY);
+      Geocode.setRegion('us');
       const response = await Geocode.fromLatLng(location.lat, location.lng);
       if (response.status === 'ZERO_RESULTS') {
         showError('Could not find the locality.');
@@ -79,7 +53,7 @@ function ClickableMap(props) {
         showError(`Geocoding error: ${response.status}`);
         return;
       }
-      const result = response.results[0];
+      const [result] = response.results;
 
       // Find city and state.
       let locality = null;
@@ -102,7 +76,7 @@ function ClickableMap(props) {
       // console.log(result);
       // console.log(locality, administrative_area_level_1);
 
-      props.onLocationChange({
+      onLocationChange({
         generalLocation: scrambledLocation,
         generalLocationName: locationName,
         preciseLocation: {
@@ -112,18 +86,20 @@ function ClickableMap(props) {
       });
       setMarkerLocation(location);
     } catch (err) {
-      showError(err);
+      console.error(`Error detecting locality: ${err.message}`, err); // eslint-disable-line no-console
+      showError(err.message);
+      throw err;
     }
-  };
+  }
 
-  const handleLocationClick = (args) => {
+  function handleLocationClick(args) {
     const newLoc = { lat: args.latLng.lat(), lng: args.latLng.lng() };
     // console.log("handle click", args);
     // console.log("handle click", newLoc);
     setLocation(newLoc);
-  };
+  }
 
-  const handleDetectLocation = () => {
+  function handleDetectLocation() {
     setDetectingLocation(true);
     if (!navigator.geolocation) {
       showError(
@@ -150,7 +126,7 @@ function ClickableMap(props) {
         setDetectingLocation(false);
       },
     );
-  };
+  }
 
   return (
     <>
